ashish check round 1 - edit
this is the planned edits from the changes made as per the meeting with ashish on feb 5th; locations zooming/lat/long changed (in script), the order of some events was changed (E1047 placed after E1048) and pictures were added (E1054 & E1062) (in HTML). at this point HTML checking by ashish has been done and we're hoping for a 2nd ashish check on feb 6th.