ashish check round 1 - edits
this is the planned edits from the changes made as per the meeting with ashish on feb 5th; locations zooming/lat/long changed (in script). at this point HTML checking by ashish has been done + sumandro has to add them. we're hoping for a 2nd ashish check on feb 6th.